requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
pydantic==2.11.7
langchain==0.3.27
langchain-community==0.3.29
langchain-openai==0.3.32
langchain-text-splitters==0.3.9
faiss-cpu==1.12.0
streamlit==1.48.1
tiktoken==0.11.0
python-multipart==0.0.20
pypdf==5.6.0
httpx==0.28.1
